image_architect/requirements.txt
google-adk>=0.4.0
google-genai>=0.8.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
a2a-sdk>=0.3.0
google-auth>=2.0.0
httpx>=0.27.0